+++ app/src/layouts/TabsLayout/index.ts
@@ -1,3 +1,3 @@
-export { TabsLayoutProvider, useTabsLayoutContext, TabsLayoutContextInterface } from "./contexts";
 export { TabsLayoutContainer } from "./container";
+export * from "./contexts";
 export * from "./types";
