+++ app/src/layouts/TabsLayout/types.ts
@@ -15,3 +15,11 @@ export namespace TabsLayout {
     timeStamp: number;
   }
 }
+
+export interface TabsLayoutContextInterface {
+  tabs: TabsLayout.Tab[];
+  activeTab: TabsLayout.Tab;
+  closeTab: (tabId: TabsLayout.Tab["id"]) => void;
+  openTab: (tabId: TabsLayout.Tab["id"], tabDetails?: Partial<TabsLayout.Tab>) => void;
+  updateTab: (tabId: TabsLayout.Tab["id"], updatedTabData?: Partial<TabsLayout.Tab>) => void;
+}
